requirements.txt
tqdm==4.42.1
wandb==0.9.7
numpy==1.18.1
matplotlib==3.1.3
Pillow==7.0.0
scikit-image==0.16.2
pycocotools==2.0.0
jsonlines==1.2.0
torch==1.5.0
torchvision==0.6.0
rasterio==1.2.8